third_party/WebKit/LayoutTests/fast/html/imports/sub-imports-loop.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<link id="root" rel="import" href="resources/loop-root.html">
</head>
<body>
<script>
rootImport = root.import;
childImport = rootImport.querySelector("link").import;
rootImportFromChild = childImport.querySelector("link").import;
shouldBe("rootImport", "rootImportFromChild");
</script>
</body>
</html>
